backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
email-validator==2.1.0
discord-webhook==1.3.1
pywebpush==1.14.0
py-vapid==1.9.0
APScheduler==3.10.4
pytz==2024.1
